enable pedantic

+++ src/game/state.rs
@@ -85,7 +85,7 @@ pub struct Game {
   pub balls: Vec<Ball>,
   pub is_game_over: bool,
   pub selected_ball: Option<SelectedBall>,
-  pub points: u32,
+  pub points: usize,
   pub can_add_balls: bool,
   pub navigation_height: i32,
   pub new_game_button: Button,
